Oprydning i kode samt tilføjelse af verifyPassword metode

+++ src/Betalinger.java
@@ -2,9 +2,9 @@ import java.util.ArrayList;
 
 public class Betalinger{
     Customer c;
-    int payment;
+    double payment;
 
-    public Betalinger(Customer c, int payment){
+    public Betalinger(Customer c, double payment){
         this.c = c;
         this.payment = payment;
     }
